# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/question-inputs/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CreateQuestionInput",
                "tags": [
                    "questionInput"
                ],
                "summary": "CreateQuestionInput",
                "parameters": [
                    {
                        "description": "questionInput",
                        "name": "questionInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.CreateQuestionInputRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.QuestionInputId"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-inputs/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "DeleteQuestionInput",
                "tags": [
                    "questionInput"
                ],
                "summary": "DeleteQuestionInput",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question input deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-inputs/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetQuestionInputById",
                "tags": [
                    "questionInput"
                ],
                "summary": "GetQuestionInputById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.GetQuestionInputResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/question-inputs/{question_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetQuestionInputsByQuestionId",
                "tags": [
                    "questionInput"
                ],
                "summary": "GetQuestionInputsByQuestionId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question_id",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.GetAllQuestionInputsByQuestionIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/test-cases/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new test case",
                "tags": [
                    "testCase"
                ],
                "summary": "CreateTestCase",
                "parameters": [
                    {
                        "description": "Create test case request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.CreateTestCaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test case created successfully",
                        "schema": {
                            "$ref": "#/definitions/question.TestCaseId"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/test-cases/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete test case",
                "tags": [
                    "testCase"
                ],
                "summary": "DeleteTestCase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Test case id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test case deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/test-cases/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get test case by id",
                "tags": [
                    "testCase"
                ],
                "summary": "GetTestCaseById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Test case id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test case retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/question.GetTestCaseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/test-cases/{question_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get test cases by question id",
                "tags": [
                    "testCase"
                ],
                "summary": "GetTestCasesByQuestionId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question id",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test cases retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/question.GetAllTestCasesByQuestionIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "question.CreateQuestionInputRequest": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "question.CreateTestCaseRequest": {
            "type": "object",
            "properties": {
                "case": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
        "question.GetAllQuestionInputsByQuestionIdResponse": {
            "type": "object",
            "properties": {
                "question_inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.GetQuestionInputResponse"
                    }
                }
            }
        },
...
        "question.GetAllTestCasesByQuestionIdResponse": {
            "type": "object",
            "properties": {
                "test_cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/question.GetTestCaseResponse"
                    }
                }
            }
        },
        "question.GetQuestionInputResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "question.GetTestCaseResponse": {
            "type": "object",
            "properties": {
                "case": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "question.QuestionInputId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "question.TestCaseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Gateway",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}